get single offer by id

+++ src/main/scala/de/hpi/epic/pricewars/Offer.scala
@@ -6,7 +6,7 @@ import spray.json._
 /**
   * Created by Jan on 01.11.2016.
   */
-case class Offer ( offer_id: Option[Int],
+case class Offer ( offer_id: Option[Long],
                    product_id: String,
                    seller_id: String,
                    amount: Int,
